pmic_lld: misrac fixes rule-R.10.3
misrac fixes rule-R.10.3 - MISRA.ETYPE.ASSIGN.2012
An expression value of essential type unsigned long int is assigned to an object of essential type unsigned char

+++ src/pmic_rtc.c
@@ -1856,7 +1856,7 @@ int32_t  Pmic_rtcEnableRtc(Pmic_CoreHandle_t *pPmicCoreHandle,
         if(PMIC_RTC_STOP == enableRtc)
         {
             /* Stopping the RTC */
-            regData &= ~(PMIC_RTC_CTRL_1_STOP_RTC_MASK);
+            regData &= (uint8_t)(~(PMIC_RTC_CTRL_1_STOP_RTC_MASK));
         }
         else
         {
